removed temp shoot sequence shuffleboard tuning

+++ RobotCode2020New/src/main/java/frc/robot/commands/hoppercommands/RunHopperCommand.java
@@ -26,7 +26,7 @@ public class RunHopperCommand extends CommandBase {
     private JoystickButton zr;
     private double counter;
 
-    private final double COUNTER_LIMIT = 1;
+    private final int COUNTER_LIMIT = 1;
     
     //-------- CONSTRUCTOR --------\\
